rpki validation: parse in batches and handle concatenated json
The rpki-client '-f' flag can handle multiple ROA files. Instead of
passing files one by one, pass them in batches of 250.
Instead of building a list of `json_results` in memory, write to
the result file as they return in batches.
This requires adding a delimiter to the resulting JSON stream, which
does not have any by default.
Finally, this removes the parallelization via ThreadPoolExecutor, and is
about 2x faster without it. May want to bring this back.